+++ AmazonChimeSDK/AmazonChimeSDK/audiovideo/DefaultAudioVideoFacade.swift
@@ -99,6 +99,14 @@ import Foundation
         try realtimeController.realtimeSendDataMessage(topic: topic, data: data, lifetimeMs: lifetimeMs)
     }
 
+    public func realtimeSetVoiceFocusEnabled(enabled: Bool) -> Bool {
+        return realtimeController.realtimeSetVoiceFocusEnabled(enabled: enabled)
+    }
+
+    public func realtimeIsVoiceFocusEnabled() -> Bool {
+        return realtimeController.realtimeIsVoiceFocusEnabled()
+    }
+
     public func addAudioVideoObserver(observer: AudioVideoObserver) {
         audioVideoController.addAudioVideoObserver(observer: observer)
     }